Let apt.postgresql.org.sh install Postgres
This allows prerelease versions to work (as long as `apt.postgresql.org.sh` is
up-to-date with its hard-coding of the pre-release, which happens once a
year).

Also, install some additional dependencies in the Docker image, and purge
postgresql-client-common, so that pg-start does not have to spend the time
doing it.