Add user customization and release on tag
Add an entrypoint script that will optionally create an unprivileged user based
on an environment variable. The UID for the user will either be the value of the
`LOCAL_UID` variable, the UID associated with a file found at
`GITHUB_EVENT_PATH` (for automatic access to files mounted in a GitHub action),
or 1001 (the currently-known GitHub runner UID).

This feature will be useful in situations when a non-superuser is required, but
also add `sudo` and  allow its use without a password, and update `pg-start` and
`pg-build-test` to use it.

Add a test to the GitHub workflow to make sure things work for an unprivileged
user as they do for root, and give all the workflow steps descriptive names.

Cease to release on every push to main, and only push on tags starting with "v".

Modify `pg-start` to show a more useful error message if no Postgres version is
passed to it.