Fix Git archiving and improve docs
In GitHub actions `pgxn-bundle` was failing to run `git archive` because
the `git rev-parse` command it uses to test whether it's working in a
Git tree was failing with the suppressed error "detected dubious
ownership in repository". Fix it by running this command in
`entrypoint.sh`:

```shell
git config --system --add safe.directory "$PWD"
```

So that the working directory is always considered safe. Also tweak the
test to fail if it doesn't use `git archive` to zip the files and add a
test to ensure it properly falls back on `zip` when not working in a Git
project. Fixes #5.

Relatedly, update `pgxn-bundle` to print whether it's archiving from a
Git repo or zipping all the files, and turn on verbosity for `git
archive` to emit the list of files added to the zip archive.

Other changes:

*   Use the Docker `--volume` option instead of `--mount` in tests and
    documentation.
*   Upgrade checkout and Docker GitHub actions.
*   Update copyright year.
*   Document installed Apt packages.