Document unprivileged user in GitHub workflows
At some point GitHub changed things so that the entrypoint script does
not run, so the `AS_USER` variable does nothing. Which is just as well,
because that user would have no permissions, so even a checkout will
fail.

So remove the GitHib-specific stuff from `entrypoint.sh` and separately
document running the container with the CLI and as a GitHub workflow.
Add a new section describing how to use `gosu` directly in a GitHub
workflow to execute commands as an unprivileged user.

Thanks to @pgguru for figuring out the technique.