fix: conditionally execute this without exit code causing issues
The original construct would return 1 if the `[` test failed, which due to the
`set -eu` earlier in the script would result in the whole script failing.
Change to use an `if` block which does not set the exit code accordingly, and
also avoids swallowing any error from `pg_createcluster` (vs using the `|| true`
approach).